requirements.txt
asgiref==3.4.1
Django==3.2.16
django-filter==21.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.8.3
Markdown==3.3.7
Pillow==8.4.0
PyJWT==2.4.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.1.1
zipp==3.6.0
